+++ src/Game/LevelPlayer/Phaser/LevelPlayer.ts
@@ -1,6 +1,5 @@
 import * as Phaser from 'phaser';
 
-import { sessionCookieValue } from '../../../../public/js/login.js';
 import config from '../../config.js';
 import ArticodingSprite from './Classes/ArticodingSprite.js';
 import ChestObject from './Classes/ChestObject.js';
@@ -265,38 +264,7 @@ export default class LevelPlayer extends Phaser.Scene {
       document.dispatchEvent(event);
     } else {
       const event = new CustomEvent("win", { detail: { stars: 3 } });
-      document.dispatchEvent(event);
-
-      const cookie = sessionCookieValue();
-      if(cookie !== null){
-        const urlParams = new URLSearchParams(window.location.search);
-        const levelId = urlParams.get('id');
-        const postData = {
-          user: cookie.id,
-          level: levelId,
-          stars: 1,
-          attempts: 1,
-        };
-
-        fetch(config.API_ENDPOINT + "/play", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify(postData),
-        })
-        .then(response => {
-          if (response.ok) {
-            console.log("La petición se ha realizado correctamente.");
-          } else {
-            console.error("La petición falló con estado:", response.status);
-          }
-        })
-        .catch(error => {
-          console.error("Error al realizar la petición:", error);
-        });
-      }
-     
+      document.dispatchEvent(event);     
     }
   }
